services/converter/pyproject.toml
[project]
name = "converter"
version = "0.1.0"
description = "API convert files for microservices"
requires-python = ">=3.10"
dependencies = [
    "flask==3.1.0",
    "flask-pymongo>=3.0.1",
    "jedi>=0.19.2",
    "moviepy>=2.2.1",
    "pika>=1.3.2",
    "pylint>=3.3.6",
    "pymongo>=4.13.2",
    "python-dotenv==1.1.0",
    "requests>=2.32.4",
]

[tool.uv]
dev-dependencies = [
    "pylint==3.3.6",
    "isort==6.0.1"
]
